Moved/backed up current engagement notif implementation in preparation for vite-migration merge.
